+++ build.gradle
@@ -28,10 +28,15 @@ repositories {
 
 }
 
+
+
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	//thymeleaf
 	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+    //Validation
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+
 
 	//便利機能、スターゆーてぃりティー
 	implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.14.0'
