+++ src/main/java/raiseteach/StudentManagement/StudentManagementApplication.java
@@ -1,13 +1,9 @@
 package raiseteach.StudentManagement;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.*;
 
-import raiseteach.StudentManagement.data.StudentCourse;
-
-import java.util.List;
+import raiseteach.StudentManagement.repository.StudentRepository;
 
 @SpringBootApplication
