+++ src/main/java/raiseteach/StudentManagement/domain/StudentDetail.java
@@ -1,13 +1,14 @@
 
 package raiseteach.StudentManagement.domain;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import raiseteach.StudentManagement.data.StudentCourse;
 import raiseteach.StudentManagement.data.StudentFolder;
-
+import org.springframework.validation.annotation.Validated;
 import java.util.List;
 
 @Getter
@@ -17,10 +18,14 @@ import java.util.List;
 public class StudentDetail {
 
     // 単一の StudentFolder を保持する
+    @Valid
     private StudentFolder studentFolder;
 
     // 複数の StudentCourse を保持するリスト
+    @Valid
     private List<StudentCourse> studentCourse;
+
+
 
     // Lombok が自動で getter と setter を生成するため、以下は不要
