+++ src/main/resources/application.properties
@@ -7,5 +7,5 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #mybatis???
 
-mybatis.configration.map-underscore-to-camel-case=true
-mybatis.mapper-locations=classpath*:/mappers/*.xml
+mybatis.configuration.map-underscore-to-camel-case=true
+mybatis.mapper-locations=classpath*:/mapper/*.xml
